Create modelwithdatacleaning
This is the model with data cleaning.  It predicts much more accurately than the models without it.  This folder contains the train_model file which trains the model using 0-58000 images.  It generates a model that can be saved and used in the predictdrivingangles file.  The predictdrivingangles generates two videos of the road and the predicted angle of the steering wheel.  This folder also contains a model that I trained if you do not have access to a GPU.